optimize requires

+++ src/main/java/com/hibegin/common/util/EnvKit.java
@@ -1,8 +1,6 @@
 package com.hibegin.common.util;
 
 import java.io.File;
-import java.lang.management.ManagementFactory;
-import java.lang.management.RuntimeMXBean;
 
 public class EnvKit {
 
@@ -20,14 +18,15 @@ public class EnvKit {
     }
 
     public static void savePid(String pidFile) {
-        RuntimeMXBean runtimeMXBean = ManagementFactory.getRuntimeMXBean();
-        int pid = Integer.valueOf(runtimeMXBean.getName().split("@")[0]);
         File file = new File(pidFile);
         if (file.exists()) {
             file.delete();
         }
-        IOUtil.writeStrToFile(pid + "", file);
-        file.deleteOnExit();
+        long pid = Pid.get();
+        if (pid > 0) {
+            IOUtil.writeStrToFile(pid + "", file);
+            file.deleteOnExit();
+        }
     }
 
     public static boolean isAndroid() {
